Finished work on the sign in form
Made frontend validation messages appear properly, according to the state of the form.

Used absolute positioning with margin-top because I did not want the form elements to move with the appearance and disapearance of validation messages. So I had to rewrite the CSS quite a bit.

The sign in form now looks good on a variety of screen sizes

Restricted the max length of username to 20 characters on the frontend yet to do server side validation.