Work on add a new picture form
---------------------------------------------------------------------

Made the mouse a cursor over both the radio input and its label to better tell the user that clicking the label has the same effect as clicking the radio input

Removed the multipleUpload boolean in the JS which is no longer being used

I made readFileIntoAURL insert the src of the picture in the previewSrc array at same index as its index in the file list.
This is important because the pictureTitle[i] field (which the server interprets and where i refers to the index of the file in the file list) and the preview image of the picture must correspond to the same file.
Before we were just pushing the src into the array. However, we cannot assume that the fileReader would read all the
files in the same length of time resulting in the correct order.

Added a loading icon for image previews that are still loading

Finally forbid picture uploads of greater than 5 pictures on the frontend

Added frontend validation to the add a new picture form

If only the picture titles are invalid added a dialog asking the user to confirm if he would like the server to generate automatic names for pictures without a picture title

---------------------------------------------------------------------------

Work on the backend
-----------------------------------------------------------------------------
I made the pictureTitle optional by allowing an empty string